pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gacli"
version = "0.1.0"
description = "Simple Google Analytics 4 CLI"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "google-analytics-data>=0.18.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
]

[project.scripts]
gacli = "gacli.cli:main"
